Added csproj and removed excel personal info.
