+++ examples/ExampleReadAndWrite/ExampleReadAndWrite.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29512.175
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleReadAndWrite", "ExampleReadAndWrite\ExampleReadAndWrite.csproj", "{392D473C-34C3-41C7-9D5A-B29B4AF80D74}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Excel.IO", "..\..\src\Excel.IO\Excel.IO.csproj", "{7F5625C6-B879-4B46-9C1F-D0319C7A7E28}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{392D473C-34C3-41C7-9D5A-B29B4AF80D74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{392D473C-34C3-41C7-9D5A-B29B4AF80D74}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{392D473C-34C3-41C7-9D5A-B29B4AF80D74}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{392D473C-34C3-41C7-9D5A-B29B4AF80D74}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7F5625C6-B879-4B46-9C1F-D0319C7A7E28}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7F5625C6-B879-4B46-9C1F-D0319C7A7E28}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7F5625C6-B879-4B46-9C1F-D0319C7A7E28}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7F5625C6-B879-4B46-9C1F-D0319C7A7E28}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {662AF04F-F976-45AE-BB8F-D3E3B8EAFA63}
+	EndGlobalSection
+EndGlobal
